modify default route instead of using rule-based routing
rule-based routing would require a rule for every docker subnet, so it's quite brittle if the docker config changes. The regular routing table already includes specific routes to the docker subnet, and we don't need to worry about tools like NetworkManager modifying the route out from under us because we're in a container.